small improvement in a method and removing comments

+++ src/main/java/algorithms/misc/MiscMath0.java
@@ -742,11 +742,6 @@ public class MiscMath0 {
     
     public static float[] calcQuartiles(int[] a, boolean isSorted) {
         
-        // 0 : 0.25
-        // 0 : 0.5
-        // 0 : 0.75
-        // 0 : 1
-        
         if (!isSorted) {
             a = Arrays.copyOf(a, a.length);
             Arrays.sort(a);
@@ -757,7 +752,7 @@ public class MiscMath0 {
         float t50 = (float)n/2.f;
         float t75 = (float)n*3.f/4.f;
         
-        int norm = 0;
+        float norm = 0;
         float[] sums = new float[4];    
         for (int i = 0; i < n; ++i) {
             if (i < t25) {
@@ -773,18 +768,13 @@ public class MiscMath0 {
         }
         
         for (int i = 0; i < sums.length; ++i) {
-            sums[i] = (float)sums[i]/(float)norm;
+            sums[i] = (float)sums[i]/norm;
         }
         
         return sums;
     }
     
     public static float[] calcQuartiles(float[] a, boolean isSorted) {
-        
-        // 0 : 0.25
-        // 0 : 0.5
-        // 0 : 0.75
-        // 0 : 1
         
         if (!isSorted) {
             a = Arrays.copyOf(a, a.length);
@@ -819,11 +809,6 @@ public class MiscMath0 {
     }
     
     public static double[] calcQuartiles(double[] a, boolean isSorted) {
-        
-        // 0 : 0.25
-        // 0 : 0.5
-        // 0 : 0.75
-        // 0 : 1
         
         if (!isSorted) {
             a = Arrays.copyOf(a, a.length);
